requirements.txt
# Core Data Science
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.2
transformers==4.36.2

# Financial Data
yfinance==0.2.28
alpha-vantage==2.3.1
pandas-datareader==0.10.0

# Web Framework
streamlit==1.29.0
fastapi==0.104.1
uvicorn==0.24.0

# Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Database
sqlalchemy==2.0.23
sqlite3

# API and HTTP
requests==2.31.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
loguru==0.7.2

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
jupyter==1.0.0

# Time Series
statsmodels==0.14.0
arch==6.2.0

# Portfolio Optimization
cvxpy==1.4.1
pypfopt==1.5.5

# Sentiment Analysis
textblob==0.17.1
vaderSentiment==3.3.2
nltk==3.8.1
